fix(client): align P4 OpenAPI schemas with Katana official API (#268)
* fix(client): align P4 OpenAPI schemas with Katana official API

Comprehensive alignment of OpenAPI spec with Katana's official API
(2026-01-20 spec), validated by integration tests against 40 live
endpoints with zero schema gaps remaining.

Schema fixes:
- Add /demand_forecasts endpoints (GET/POST/DELETE)
- Replace SalesOrderFulfillment schema with actual API fields
- Fix SalesReturnRow fields (add fulfillment_row_id, reason_id, etc.)
- Fix StockAdjustment fields (adjustment_date -> stock_adjustment_date)
- Fix StockTransfer fields (add order_created_date, expected_arrival_date)
- Fix PurchaseOrderAccountingMetadata camelCase -> snake_case
- Fix ProductOperationRow schema (uses product_operation_row_id, not id)
- Make nullable fields explicit (PurchaseOrderAdditionalCostRow,
  SerialNumber transaction_date, etc.)
- Add missing fields to 10+ schemas (BomRow.rank, PriceList.is_active,
  Inventory.serial_tracked, etc.)

Entity inheritance fixes:
- Promote 7 schemas to DeletableEntity (SalesOrder, SalesOrderRow,
  SalesOrderAddress, CustomerAddress, PurchaseOrderRow, Stocktake,
  CustomFieldsCollection)
- Remove duplicate id fields from 5 schemas using entity composition

Response wrapping fixes:
- Add OperatorListResponse and SerialNumberStockListResponse wrappers
- Fix pagination transport for list responses that return raw arrays

Infrastructure:
- Update extract script for new README.io page structure
- Add GET endpoint schema validation integration tests
- Update comprehensive spec to 2026-01-20 version


* docs: update endpoint and model counts after P4 alignment

Update stale references from "76+ API endpoints" to "100+" and
"150+ data models" to "370+" across CLAUDE.md, architecture quick
reference, and copilot instructions. Add demand forecasting to the
API coverage categories.


* fix(mcp): update stock adjustment field references after P4 alignment

StockAdjustment schema changes in P4:
- adjustment_date renamed to stock_adjustment_date
- status and reference_no fields removed
- Replace reference_no with reason (actual API field)


* fix(client): regenerate TypeScript client for P4 schema alignment

Regenerate TS types and SDK from updated OpenAPI spec. Adds demand
forecast endpoints, fixes field names across 20+ schemas, and syncs
all P4 changes (entity inheritance, nullable fields, response
wrapping).


* fix(client): address PR review comments

- Fix LIST_ENDPOINTS comment to match actual 3-tuple shape
- Rename TestEnumDefaults to TestRequestModels (class no longer tests enums)
- Add explanatory comment to silent except clause in extract script
- Remove redundant asyncio import in test_real_api_pagination


* docs: update endpoint count in OpenAPI docs page


* fix(client): address second round of PR review comments

- Fix MRO generator to handle both BaseEntity orderings in class
  inheritance (was only removing BaseEntity when it appeared first)
- Make schema validation test delay configurable via SCHEMA_TEST_DELAY
  env var (default 1.1s) to allow faster runs in controlled environments


* docs: add Katana API questions document for P4 alignment follow-up

Tracks 9 questions across 5 categories (schema asymmetries, naming
inconsistencies, missing write endpoints, nullable semantics, and
non-standard patterns) discovered during P1-P4 spec alignment, plus a
resolved-issues reference table.


* docs: add live API investigation findings to API questions document

Tested all 9 questions against the live Katana API. Key findings:
- Material serial_tracked/operations_in_sequence confirmed read-only
- MaterialConfig id/material_id on create confirmed as spec error
- MO-SO linking resolved via /manufacturing_order_make_to_order
- PO status on create confirmed pointless but harmless
- ProductOperationRow non-standard ID confirmed as real inconsistency
- Read-only endpoints (additional_costs, factory) confirmed API-only
- Null semantics clarified as "not set" (distinct from zero)
- demand_forecasts confirmed as computation endpoint, not CRUD


* fix(client): preserve original response shape in PaginationTransport

PaginationTransport was wrapping all responses in {"data": [...]}
even for non-paginated endpoints and raw JSON list responses. This
broke generated parsers that expect the original response format.

Two changes:
- Non-paginated responses (no pagination headers/metadata) are now
  returned as-is instead of being re-synthesized
- Multi-page raw list responses preserve the list format instead of
  wrapping in {"data": [...]}


* refactor(client): extract 22 inline schemas to named schemas in OpenAPI spec

Replace inline request/response body schemas with named $ref schemas
in components/schemas for consistent type generation across Python and
TypeScript clients. Anonymous types like `UpdateCustomerAddressBody` and
`GetSalesReturnReasonsResponse200Item` become properly named
`UpdateCustomerAddressRequest` and `SalesReturnReason`.

- Extract 16 request body schemas (Create*/Update*Request)
- Extract 6 response body schemas (ListResponse, entity types)
- Extract 5 nested inline schemas (BatchTransactionRequest, etc.)
- Update DOMAIN_GROUPS to fix circular imports from new schema names
- Regenerate Python client, pydantic models, and TypeScript client


* fix(client): address PR review comments

- Change ErrorResponse.statusCode from type: number to type: integer
  in OpenAPI spec so generated models use int instead of float
- Apply max_items truncation in PaginationTransport even when no
  pagination info is detected (single-page responses)
- Improve schema test delay comment to document SCHEMA_TEST_DELAY
  env var and explain the 1.1s default (Katana rate limit)


* fix: address second round of PR review comments

- Remove hard-coded 100000 char size guard in extract_all_katana_docs.py
- Replace brittle `response is mock_response` identity assertions with
  content-based assertions in transport tests
- Add missing `status` field to adjustment movements for schema consistency


---------

+++ katana_public_api_client/models/error_response.py
@@ -17,7 +17,7 @@ T = TypeVar("T", bound="ErrorResponse")
 class ErrorResponse:
     """Base error message schema"""
 
-    status_code: float | Unset = UNSET
+    status_code: int | Unset = UNSET
     name: str | Unset = UNSET
     message: str | Unset = UNSET
     additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
